add MPTCP socket protocol (optional)
Add the possibility to use the MPTCP protocol at the
socket level for users of ServerBuilder.

MPTCP is now more widely available since Linux Kernel
version >= 5.6. But it still need to be enabled
manually using: `sysctl net.mptcp.enabled=1`.
(of course, MPTCP is only available on Linux)

The new MPTCP struct give the user the option to
determine how we'll handle the case where MPTCP is
not available on the host, either we crash, or we
fallback to regular TCP.

Signed-off-by: Martin Andre <martin.andre@tessares.net>